+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # Changelog
 
+## [6.0.1-beta.4] 2025-11-14
+## Changed
+- Require Python 3.13
+
+## [6.0.1-beta.3] 2025-11-13
+## Added
+- Support sandboxed flatpak
+
+## Changed
+- Preferred client from radio toggle to combobox
+
+## Fixed
+- Config file erroneously getting updated when populating settings menu
+
+## [6.0.1-beta.2] 2025-11-10
+## Fixed
+- Python 3.14 error in finally block
+- Require Python version between 3.11 and 3.12
+
+## [6.0.1-beta.1] 2025-11-05
+## Fixed
+- UI not being constructed correctly if CHANGELOG.md was missing
+
 ## [6.0.0-beta.8] 2025-11-04
 ## Added
 - Internal flag to allow distribution packaged releases to disable in-app updates